backend/requirements.txt
Flask==3.0.2
pymongo[srv]==4.7.0
python-dotenv==1.0.1
flask-cors==4.0.0

